+++ Transceiver52M/arch/common/convolve_base.c
@@ -146,7 +146,7 @@ int base_convolve_complex(const float *x, int x_len,
 }
 
 /* Aligned filter tap allocation */
-void *convolve_h_alloc(int len)
+void *convolve_h_alloc(size_t len)
 {
 #ifdef HAVE_SSE3
 	return memalign(16, len * 2 * sizeof(float));
